requirements.txt.txt
torch>=1.9.0
torchvision>=0.10.0
streamlit>=1.28.0
Pillow>=8.0.0
numpy>=1.21.0
opencv-python-headless>=4.5.0